test(vaults): deposit/withdraw in the same tx
