+++ test/vaults/utils.ts
@@ -7,8 +7,11 @@ import ProofsFetcher from "../../src/vaults/ProofsFetcher";
 
 export const distributorFromEvents = (vaultAddress: string, events: TransactionEvents[]): Distributor => {
   class EventFetcherOneDepositor implements EventsFetcherInterface {
+    /* eslint-disable-next-line */
     async fetchSortedEventsForEpoch(epochConfig: EpochConfig): Promise<[TransactionEvents[], BigNumber]> {
-      return [events, epochConfig.initialTimestamp];
+      const blockFrom = events[0].event.blockNumber;
+      const timeFrom = await this.getBlock(blockFrom).then((block) => BigNumber.from(block.timestamp));
+      return [events, timeFrom];
     }
     async getBlock(blockNumber: number): Promise<providers.Block> {
       const provider = new providers.JsonRpcProvider(process.env.RPC_URL);
